Split bump internals tests into four cohesive modules (#97)
Pure test reorganisation — no test logic, assertions, or Hypothesis
strategies changed.

- test_bump_command_internals.py: retains the selector/skip, change-count,
  and result-message tests, plus the shared `UpdateCrateTestParams` and
  workspace-construction helpers.
- test_bump_manifest_editing.py (new): crate-manifest update and low-level
  TOML table/version helpers, with `_parse_manifest_versions`. Imports the
  shared params/builder from test_bump_command_internals rather than
  duplicating them.
- test_bump_dependency_sections.py (new): the four
  `_update_dependency_sections` workspace-flag tests.
- test_bump_context_properties.py (new): the issue #97 property tests and
  their synthetic-workspace / manifest helpers.

Each new module carries `from __future__ import annotations` and only the
imports its tests use. All 29 test functions (30 items) collect and pass with
identical names; make check-fmt/lint/typecheck all green.
